+++ IntroductionToAlgorithms.Tests/Sorting/QuickSortTests.cs
@@ -12,9 +12,9 @@ namespace IntroductionToAlgorithms.Tests.Sorting
     [TestClass]
     public class QuickSortTests : SortTests
     {
-        protected override ISort<int> CreateSortAlgorithm()
+        protected override ISort<byte> CreateSortAlgorithm()
         {
-            return new QuickSort<int>();
+            return new QuickSort<byte>();
         }
     }
 }
